requirements.txt
fastapi
uvicorn[standard]
google-generativeai
supabase
python-dotenv